antsbimbel-be/requirements.txt
Django>=5.2
djangorestframework>=3.15
drf-spectacular>=0.27
Pillow>=10.0
